Entities, Components displayed in a tree. UI enhanced.
Entities and components are now displayed in a tree view.
Each can be selected.

For now, a selected entity / component just changes its color,
it could be nicer to have a glow or something like this.

Next step is to display components drawables and different informations about
materials, transformation, etc.