Show/hide gizmos instead of delete/creating them
This commit is the change that was needed for a long time. Before, each
time the selected item changed, the current gizmo was deleted and a new
one was created. This led to very tricky bugs with cascade deletions of
render objects.

This is now simpler : the gizmos are created at startup of the gizmo
manager and simply shown or hidden.