Solved the problem on retina screen and, hopefully, on all the exotic screens one might use somewhere ... This solved the "retina problem" (picture only 1/4 of the size of the window) by doing 2 things :    - Remove the dependancy on QOpenGLWidget : the Viewer class no more inherit from QOpenGLWidget. This has also the beneficial side effect to removing OpenGL stuff include headache.    - Remove all the erroneous glViewport stuff that were spread anywhere in the code (hopefully removed all). glViewport is a global state in OpenGL not a per FBO state nor a per draw call nor ...
Compare to the last tentative, no specific code was introduced (no x2 on FBO, no use of pixelratio ...) Just a correction of 1 bug spread across the whole engine and removing of 1 non needed dependency.

That being done, Radium work now as fine as possible, without performance problems, on MacOsX with or without retina (or even with both).