Fix crash when loading models with main-app
LightManager was calling m_data->clear(), causing partial deletion of lights.
As a result, the method lightManager::count() was returning 0 after loading a file, while the camera_light was still in the manager component storage.
At runtime, this raise a CORE_ASSERT when trying to add the camera_light in Gui::Viewer::startRendering.