Add multiple modifiers support.
Scale whole is now part of the keyMapping.
But maybe a cleaner solution to not pollute other gizmo may be better.

Add error message when context is invalid.

Also remove the core assert so the application won't crash.
And add a config file that is not complete on purpose to test these behavior.

Resolve error and warning VS.

GizmoManager updateValues move to mouseEvent.

Plus some typos and consistency.