Add a mechanism to load only compatible plugins (debug/release)
The mechanism is based on QtPluginsMetadata, which allows to pass informations to the loader without loading the library.
We have to do this as the plugins links to the same libraries than radium, and thus conflicts arise when the build modes are different between the application and the plugin.
We cannot simply load, check and unload the plugin, as this acutally replace all the conflicting libraries.